Minor fixes in a view of c#-porting.
DEVSIX-1646

Autoported commit.
Original commit hash: [8d8266f94]

+++ itext/itext.layout/itext/layout/renderer/TableWidths.cs
@@ -580,7 +580,7 @@ namespace iText.Layout.Renderer {
                 }
                 else {
                     if (!GetTable().IsComplete() && null != GetTable().GetWidth() && GetTable().GetWidth().IsPercentValue()) {
-                        GetTable().SetWidth(tableRenderer.RetrieveUnitValue(availableWidth, Property.WIDTH));
+                        GetTable().SetWidth((float)tableRenderer.RetrieveUnitValue(availableWidth, Property.WIDTH));
                     }
                 }
                 fixedTableWidth = true;
